Updated branch name handling in `sonarParams` list
Updated the `sonarParams` list to include the branch name as a parameter. Split the condition checking if the branch name is "develop" or "master" into two separate conditions. Removed redundant addition of the branch name for 'feature/', 'hotfix/', and 'release/' branches. Moved a comment to the condition where the branch name is "master".