Updated Extensions and Added Unit Tests
Updated `Extensions.RabbitMQ.cs` and `Extensions.Telemetry.cs` to improve handling of various scenarios and added a new test class `TelemetryExtensionsTests` in `TelemetryExtensionsTests.cs`.

In `Extensions.RabbitMQ.cs`, the `AsString` method has been updated to handle cases where the content is a string or a byte array. For `Extensions.Telemetry.cs`, the `SafeStartActivity` methods have been updated to create a new `Activity` with the provided name. The `SetTelemetry` method now uses a ternary operator for setting the SpanId and TraceId. A new `SetHeader` method has been added and the `SetTraceId` and `SetSpanId` methods have been updated to use this new method.

A new test class `TelemetryExtensionsTests` has been added to `TelemetryExtensionsTests.cs` to cover unit tests for the methods in `Extensions.Telemetry.cs`.